refactor(server): change llm struct to dyn trait in llm cache
