+++ apps/server/api/src/ws/llm/models/qwen3/mod.rs
@@ -1,0 +1,1 @@
+pub mod quantized;
