+++ apps/server/api/src/ws/llm/models/token_output_stream.rs
